# src/spec/openapi_0.0.1.yaml
openapi: 3.0.0
info:
  title: Control Plane - OpenAPI 3.0
  description: |-
    This is a  Control Plane API Documentation based on the OpenAPI 3.0 specification.
  version: 0.0.1
servers:
  - url: https://controlplane.test.io/api/v1
paths:
  /clusters:
    get:
      tags:
        - Clusters
      summary: Retrieve all available clusters
      security:
        - OAuth2: [admin]
      operationId: listClusters
      parameters:
        - required: false
          description: filter Condition for applying filters to the query
          schema:
            type: string
          name: filter
          in: query
      responses:
        "200":
          description: A list of available clusters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClusterListResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
    post:
      tags:
        - Clusters
      summary: Create a new cluster
      operationId: createCluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterRequest"
      responses:
        "201":
          description: The created cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "409":
          description: If the Cluster already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
  /clusters/{clusterId}:
    get:
      tags:
        - Clusters
      summary: Retrieve a specific cluster
      operationId: getCluster
      parameters:
        - name: clusterId
          in: path
          required: true
          description: The ID of the cluster to retrieve
          schema:
            type: string
      responses:
        "200":
          description: The requested cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterResponse"
        "404":
          description: The requested cluster was not found
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
  /applications:
    get:
      tags:
        - Applications
      summary: Retrieve all available Applications
      operationId: listApplications
      parameters:
        - required: false
          description: filter Condition for applying filters to the query
          schema:
            type: string
          name: filter
          in: query
      responses:
        "200":
          description: A list of available Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
    post:
      tags:
        - Applications
      summary: Create a new Applications
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationRequest"
      responses:
        "201":
          description: The created Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "400":
          description: If the request is not well formatted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "403":
          description: If the user is not authorized to perform this query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "409":
          description: If the Application already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "422":
          description: If Request Validation Fails for the Payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          x-ms-error-response: true
        "500":
          description: Internal Server Error
          x-ms-error-response: true
        "503":
          description:
            If the API service or one of its downstream dependencies is
            temporary unavailable
          x-ms-error-response: true
  /applications/{applicationId}:
    get:
      tags:
        - Applications
      summary: Retrieve a specific Application
      operationId: getApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          description: The ID of the Application to retrieve
          schema:
            type: string
      responses:
        "200":
          description: The requested Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "404":
          description: The requested Application was not found
...
components:
...
  schemas:
    Metadata:
      type: object
    ClusterRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the cluster
        description:
          type: string
          description: A description of the cluster
        short_name:
          type: string
          description: A description of the cluster
        metadata:
          $ref: "#/components/schemas/Metadata"
    ClusterResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the cluster
        name:
          type: string
          description: The name of the cluster
        description:
          type: string
          description: A description of the cluster
        short_name:
          type: string
          description: A description of the cluster
        environment:
          type: string
          description: Name of environment
        created_by:
          type: string
          description: details of user who created the cluster
        updated_by:
          type: string
          description: details of user who updated the cluster
        created_on:
          type: number
          description: created date epoch
        updated_on:
          type: number
          description: updated date epoch
        metadata:
          $ref: "#/components/schemas/Metadata"
    ClusterListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ClusterResponse"
          minItems: 1
          maxItems: 20
          description: The list of Clusters
      additionalProperties: false
      description: A wrapper for the list of Clusters
    ApplicationRequest:
      type: object
      properties:
        name:
          type: string
          description: The name of the application
        description:
          type: string
          description: A description of the application
        namespace:
          type: string
          description: The namespace of the application
        repo_url:
          type: string
          description: The repo URL of the application
        repo_branch:
          type: string
          description: The repo Branch of the application
        repo_path:
          type: string
          description: The repo Path of the application
        metadata:
          $ref: "#/components/schemas/Metadata"
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for the cluster
        name:
          type: string
          description: The name of the cluster
        description:
          type: string
          description: A description of the cluster
        namespace:
          type: string
          description: The namespace of the application
        repo_url:
          type: string
          description: The repo URL of the application
        repo_branch:
          type: string
          description: The repo Branch of the application`
        repo_path:
          type: string
          description: The repo Path of the application
        created_by:
          type: string
          description: details of user who created the cluster
        updated_by:
          type: string
          description: details of user who updated the cluster
        created_on:
          type: number
          description: created date epoch
        updated_on:
          type: number
          description: updated date epoch
        metadata:
          $ref: "#/components/schemas/Metadata"
...
    ProblemDetails:
      type: object
      properties:
        errorCode:
          description: An application-specific error code, expressed as a string value.
          type: string
        errorDescription:
          description:
            A human-readable explanation specific to this occurrence of
            the problem. Like title, this field's value can be localized.
          type: string
        errorDetails:
          type: array
          items:
            type: object
            properties:
              propertyName:
                description: name of the field in the request which has the error
                type: string
              propertyError:
                description: the error associated  property
                type: string
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://controlplane.localhost/oauth/authorize
          tokenUrl: https://controlplane.localhost/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations